+++ firmware/spade/src/shared/sprig_engine/module_native.c
@@ -9,7 +9,7 @@
 #include "native_magic_strings.h"
 
 #include "shared/js_runtime/js.h"
-#include "shared/FatFs_SPI/ff15/source/ff.h"
+#include "shared/sprig_sd/FatFs_SPI/ff15/source/ff.h"
 
 static struct {
   jerry_value_t x, y, dx, dy, addr, type, _x, _y, _type, push, remove, generation;
