+++ firmware/spade/src/shared/sprig_sd/FatFs_SPI/ff15/source/00readme.txt
