+++ firmware/spade/src/shared/sprig_sd/FatFs_SPI/sd_driver/hw_config.h
