+++ firmware/spade/src/shared/sprig_sd/FatFs_SPI/src/glue.c
