+++ firmware/spade/src/shared/sprig_sd/sd_imports.h
@@ -1,0 +1,4 @@
+#include "shared/sprig_sd/FatFs_SPI/include/f_util.h"
+#include "shared/sprig_sd/FatFs_SPI/ff15/source/ff.h"
+#include "shared/sprig_sd/FatFs_SPI/include/rtc.h"
+#include "shared/sprig_sd/hw_config.c"
